Updated readme to download data. Trying to get get liftOver code up

+++ scripts/get_all_ct_segment_folder.py
@@ -18,7 +18,7 @@ def transform_state_into_desired_form(state, state_annot_dict):
 	try: 
 		int_state = int(state[1:]) # if it is in the form such as E1, E2, ... E18
 	except:
-		try: 
+		try:
 			int_state = state_annot_dict[state] # if not in the above from, it should be of the form that are consistent with the mneunomic
 		except:
 			print("state annotation is not in the right format: {}".format(state))
@@ -39,13 +39,15 @@ def process_one_ct_one_chrom_segment_fn(fn, ct, chrom, state_annot_dict):
 		segment_series.extend([row['state']] * row['num_bin'])
 	return segment_series
 
-def print_one_chrom_data_to_text(chrom_df, chrom, output_folder):
+def print_one_chrom_data_to_text(chrom_df, chrom, output_folder, redo_existing_files):
 	# save the data from chromsome into multiple files, each corresponding to a region on the chromosome
 	num_files_to_print = int(np.ceil(chrom_df.shape[0] / helper.NUM_BIN_PER_WINDOW))
 	for file_index in range(num_files_to_print - 1):
+		save_fn = os.path.join(output_folder, chrom + '_' + str(file_index) + '_combined_segment.bed.gz')
+		if (os.path.isfile(save_fn)) and redo_existing_files == 0:  # file already exists and users specifies that they do not need to rewrite this file
+			continue 
 		start_row_index = file_index * helper.NUM_BIN_PER_WINDOW
 		end_row_index = start_row_index + helper.NUM_BIN_PER_WINDOW
-		save_fn = os.path.join(output_folder, chrom + '_' + str(file_index) + '_combined_segment.bed.gz')
 		save_df = chrom_df.loc[start_row_index:end_row_index]
 		save_df.to_csv(save_fn, header = True, index = False, sep = '\t', compression = 'gzip')
 	last_index = num_files_to_print - 1
@@ -55,7 +57,7 @@ def print_one_chrom_data_to_text(chrom_df, chrom, output_folder):
 	print("Done saving file: {}".format(last_save_fn))
 	return 
 
-def process_one_chrom_all_ct(input_fn_list, ct_list, chrom, output_folder, state_annot_dict):
+def process_one_chrom_all_ct(input_fn_list, ct_list, chrom, output_folder, state_annot_dict, redo_existing_files):
 	result_df = pd.DataFrame()
 	for ct_index, ct in enumerate(ct_list):
 		segment_series = process_one_ct_one_chrom_segment_fn(input_fn_list[ct_index], ct, chrom, state_annot_dict)
@@ -69,23 +71,22 @@ def process_one_chrom_all_ct(input_fn_list, ct_list, chrom, output_folder, state
 	result_df['start_bp_this_window'] = helper.NUM_BP_PER_BIN * result_df.index
 	result_df['end_bp_this_window'] = result_df['start_bp_this_window'] + helper.NUM_BP_PER_BIN
 	result_df = result_df[['chrom', 'start_bp_this_window', 'end_bp_this_window'] + ct_list]
-	print_one_chrom_data_to_text(result_df, chrom, output_folder)
+	print_one_chrom_data_to_text(result_df, chrom, output_folder, redo_existing_files)
 	return 
 
-def combine_all_ct_segment_folder_one_process(input_fn_list, ct_list, chrom_list, output_folder, state_annot_dict):
+def combine_all_ct_segment_folder_one_process(input_fn_list, ct_list, chrom_list, output_folder, state_annot_dict, redo_existing_files):
 	for chrom in chrom_list:
 		chrom = 'chr' + chrom # currenntly chrom is just 1 --> 22, X. Again, we exclude Y because not every samples have Ys data
-		process_one_chrom_all_ct(input_fn_list, ct_list, chrom, output_folder, state_annot_dict)
+		process_one_chrom_all_ct(input_fn_list, ct_list, chrom, output_folder, state_annot_dict, redo_existing_files)
 	return 
 
-def combine_all_ct_segment_folder(input_folder, output_folder, input_suffix, state_annot_dict):
+def combine_all_ct_segment_folder(input_folder, output_folder, input_suffix, state_annot_dict, redo_existing_files):
 	input_fn_list = glob.glob(input_folder + '/*/*' + input_suffix) # ex: input_folder/E003_chr22_core_K27ac_segments.bed.gz
-	print(input_fn_list)
 	ct_list = list(map(lambda x: x.split('/')[-1].split(input_suffix)[0], input_fn_list)) # E003
 	print (ct_list)
 	num_cores = 4
 	partition_chrom_list = helper.partition_file_list(helper.CHROMOSOME_LIST, num_cores)
-	processes = [mp.Process(target = combine_all_ct_segment_folder_one_process, args = (input_fn_list, ct_list, partition_chrom_list[i], output_folder, state_annot_dict)) for i in range(num_cores)]
+	processes = [mp.Process(target = combine_all_ct_segment_folder_one_process, args = (input_fn_list, ct_list, partition_chrom_list[i], output_folder, state_annot_dict, redo_existing_files)) for i in range(num_cores)]
 	for p in processes:
 		p.start()
 	for i, p in enumerate(processes):
@@ -94,7 +95,7 @@ def combine_all_ct_segment_folder(input_folder, output_folder, input_suffix, sta
 	return 
 
 def main():
-	if len(sys.argv) != 5:
+	if len(sys.argv) != 6:
 		usage()
 	input_folder = sys.argv[1]
 	helper.check_dir_exist(input_folder)
@@ -103,14 +104,18 @@ def main():
 	input_suffix = sys.argv[3]
 	state_annot_fn = sys.argv[4]
 	state_annot_dict = read_state_annot(state_annot_fn)
+	redo_existing_files = helper.get_command_line_integer(sys.argv[5])
+	assert redo_existing_files in [0,1], 'redo_existing_files should be 1 (yes, rewrite all the existing files in the output_folder, or 0 (no, only write files that have not been produced)'
 	print ("Done getting command line arguments")
-	combine_all_ct_segment_folder(input_folder, output_folder, input_suffix, state_annot_dict)
+	combine_all_ct_segment_folder(input_folder, output_folder, input_suffix, state_annot_dict, redo_existing_files)
 	print ('Done!')
+
 def usage():
 	print ("python get_all_ct_segment_folder.py")
 	print ("input_folder: where all the raw data for all the ct are stored")
 	print ("output_folder: where we store files that correspond to regions on the genome, and within each file, we store the data of segmentation for all the ct that are in the input_folder")
 	print ("input_suffix: inside the input_folder, the file names are <ct><input_suffix>. In testdata, it would be chr22_core_K27ac_segments.bed.gz")
 	print ('state_annot_fn: the annotation of states (state names, state index, etc.)')
+	print ('redo_existing_files: 0 or 1: 1 (yes, rewrite all the existing files in the output_folder, or 0 (no, only write files that have not been produced)')
 	exit(1)
 main()
